+++ Formula/php73-common.rb
@@ -172,10 +172,12 @@ class Php73Common < Formula
       end
     end
 
-    if nginx_snippet_file
-        nginx_config_dir.mkpath
-        File.delete nginx_config_path if File.exist?(nginx_config_path)
-        nginx_config_path.write(nginx_snippet_file)
+    if build.with? "nginx"
+      if nginx_snippet_file
+          nginx_config_dir.mkpath
+          File.delete nginx_config_path if File.exist?(nginx_config_path)
+          nginx_config_path.write(nginx_snippet_file)
+      end
     end
   end
 end
